Move the index state handling from FDBRecordStore to IndexStateManager.
  This is done mostly as a cleanup.
